:truck: [#1064] -- move template validator to project-wide validators
