+++ src/openforms/config/models.py
@@ -11,10 +11,10 @@ from solo.models import SingletonModel
 from tinymce.models import HTMLField
 
 from openforms.data_removal.constants import RemovalMethods
-from openforms.emails.validators import DjangoTemplateValidator
 from openforms.payments.validators import validate_payment_order_id_prefix
 from openforms.utils.fields import SVGOrImageField
 from openforms.utils.translations import ensure_default_language, runtime_gettext
+from openforms.utils.validators import DjangoTemplateValidator
 
 
 @ensure_default_language()
