+++ src/openforms/utils/validators.py
@@ -1,9 +1,10 @@
-from typing import List, Type
+from typing import List, Optional, Type
 
 from django.core.exceptions import ValidationError
 from django.core.validators import RegexValidator
+from django.template import Context, Template, TemplateSyntaxError
 from django.utils.deconstruct import deconstructible
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 from rest_framework import serializers
 
@@ -130,3 +131,78 @@ class SerializerValidator:
                 self.message.format(errors=serializer.errors),
                 code=self.code,
             )
+
+
+@deconstructible
+class DjangoTemplateValidator:
+    """
+    Validate template code to be a valid Django template.
+
+    This validators ensure that the template is syntactically correct. Additionally,
+    it can enforce the presence of certain required template tags.
+    """
+
+    def __init__(self, required_template_tags: Optional[List[str]] = None):
+        self.required_template_tags = required_template_tags
+
+    def __call__(self, value: str) -> None:
+        self.check_syntax_errors(value)
+        self.check_required_tags(value)
+
+    def check_required_tags(self, value):
+        if not self.required_template_tags:
+            return
+
+        # since we already checked valid syntax we'll keep it simple and use string search
+        #   instead of looking through the parsed structure
+        for tag_name in self.required_template_tags:
+            # note: the double {{ and }} are escapes for the format()
+            variants = [
+                "{{% {t} %}}".format(t=tag_name),
+                "{{% {t}%}}".format(t=tag_name),
+                "{{%{t} %}}".format(t=tag_name),
+                "{{%{t}%}}".format(t=tag_name),
+            ]
+            for tag in variants:
+                if tag in value:
+                    break
+            else:
+                tag_str = "{{% {} %}}".format(tag_name)
+                raise ValidationError(
+                    _("Missing required template-tag {tag}").format(tag=tag_str),
+                    code="invalid",
+                )
+
+    def check_syntax_errors(self, value):
+        # code lifted from maykinmedia/mail-editor
+        # https://github.com/maykinmedia/mail-editor/blob/e9ea1762af9a5c7ec0826876cb647dea444b95ba/mail_editor/mail_template.py#L28
+
+        try:
+            return Template(value)
+        except TemplateSyntaxError as exc:
+            error_tpl = """
+                <p>{{ error }}</p>
+                {% if source %}
+                    {{ source|linenumbers|linebreaks }}
+                {% endif %}
+            """
+            if hasattr(exc, "django_template_source"):
+                source = exc.django_template_source[0].source
+                pz = exc.django_template_source[1]
+                highlighted_pz = ">>>>{0}<<<<".format(source[pz[0] : pz[1]])
+                source = "{0}{1}{2}".format(
+                    source[: pz[0]], highlighted_pz, source[pz[1] :]
+                )
+                _error = _("TemplateSyntaxError: {0}").format(exc.args[0])
+            elif hasattr(exc, "template_debug"):
+                _error = _("TemplateSyntaxError: {0}").format(
+                    exc.template_debug.get("message")
+                )
+                source = "{}".format(exc.template_debug.get("during"))
+            else:
+                _error = exc
+                source = None
+            error = Template(error_tpl).render(
+                Context({"error": _error, "source": source})
+            )
+            raise ValidationError(error, code="syntax_error")
